requirements.txt
streamlit
pandas
altair
websocket-client
streamlit-autorefresh